refactor(RUN-804): rename surface tree
