--- apps/gui/src/features/runs/runTable.styles.ts
@@ -1,22 +1,0 @@
-export const RUN_TABLE_CONTAINER_CLASS = "overflow-hidden rounded-lg bg-surface-primary";
-
-export const RUN_TABLE_CLASS = "text-sm";
-
-export const RUN_TABLE_HEADER_ROW_CLASS = "hover:bg-transparent";
-
-export const RUN_TABLE_HEAD_CLASS =
-  "h-9 border-b border-border-subtle px-2.5 text-2xs font-medium uppercase tracking-wider text-muted";
-
-export const RUN_TABLE_STATUS_HEAD_CLASS =
-  "w-9 border-b border-border-subtle px-0 text-center";
-
-export const RUN_TABLE_ROW_CLASS =
-  "h-[var(--control-height-lg)] cursor-pointer hover:bg-surface-hover";
-
-export const RUN_TABLE_CELL_CLASS = "border-b-0 px-2.5 py-0";
-
-export const RUN_TABLE_STATUS_CELL_CLASS =
-  "relative w-9 border-b-0 px-0 py-0 text-center";
-
-export const RUN_TABLE_REGRESSION_STRIPE_CLASS =
-  "before:absolute before:bottom-0 before:left-0 before:top-0 before:w-[3px] before:rounded-r-sm before:bg-warning-9";
