+++ apps/gui/src/features/surface/LazyMonacoEditor.tsx
