+++ apps/gui/src/features/surface/PaletteSidebar.tsx
