+++ apps/gui/src/features/surface/SurfaceBottomPanel.tsx
@@ -4,9 +4,27 @@ import { useWorkflowRegressions } from "@/queries/workflows";
 import { useCanvasStore } from "@/store/canvas";
 import { mapSSEEventToStoreAction } from "./useRunStream";
 import { useNavigate } from "react-router";
-import { RunsTable } from "../runs/RunsTable";
+import { Badge } from "@runsight/ui/badge";
+import { cn } from "@runsight/ui/utils";
+import {
+  Table,
+  TableBody,
+  TableCell,
+  TableHead,
+  TableHeader,
+  TableRow,
+} from "@runsight/ui/table";
+import {
+  Tooltip,
+  TooltipContent,
+  TooltipProvider,
+  TooltipTrigger,
+} from "@runsight/ui/tooltip";
+import { AlertTriangle } from "lucide-react";
+import type { RunResponse } from "@runsight/shared/zod";
 import { formatRegressionTooltip } from "../workflows/regressionBadge.utils";
 import { RegressionTooltipBody } from "@/components/shared/RegressionTooltipBody";
+import { formatCost, formatDuration, getTimeAgo } from "@/utils/formatting";
 
 interface LogEntry {
   timestamp: string | number;
@@ -14,7 +32,7 @@ interface LogEntry {
   message: string;
 }
 
-interface CanvasBottomPanelProps {
+interface SurfaceBottomPanelProps {
   runId?: string;
   workflowId?: string;
   defaultState?: "collapsed" | "expanded";
@@ -29,9 +47,217 @@ type RegressionsData = {
   issues?: Array<Record<string, unknown>>;
 };
 
-type CanvasBottomPanelContentProps = CanvasBottomPanelProps & {
+type SurfaceBottomPanelContentProps = SurfaceBottomPanelProps & {
   regressionsData?: RegressionsData;
 };
+
+const SURFACE_RUNS_TABLE_CONTAINER_CLASS = "overflow-hidden rounded-lg bg-surface-primary";
+const SURFACE_RUNS_TABLE_CLASS = "text-sm";
+const SURFACE_RUNS_TABLE_HEADER_ROW_CLASS = "hover:bg-transparent";
+const SURFACE_RUNS_TABLE_HEAD_CLASS =
+  "h-9 border-b border-border-subtle px-2.5 text-2xs font-medium uppercase tracking-wider text-muted";
+const SURFACE_RUNS_TABLE_STATUS_HEAD_CLASS =
+  "w-9 border-b border-border-subtle px-0 text-center";
+const SURFACE_RUNS_TABLE_ROW_CLASS =
+  "h-[var(--control-height-lg)] cursor-pointer hover:bg-surface-hover";
+const SURFACE_RUNS_TABLE_CELL_CLASS = "border-b-0 px-2.5 py-0";
+const SURFACE_RUNS_TABLE_STATUS_CELL_CLASS =
+  "relative w-9 border-b-0 px-0 py-0 text-center";
+const SURFACE_RUNS_TABLE_REGRESSION_STRIPE_CLASS =
+  "before:absolute before:bottom-0 before:left-0 before:top-0 before:w-[3px] before:rounded-r-sm before:bg-warning-9";
+
+function getRunStatusTone(status: string) {
+  switch (status.toLowerCase()) {
+    case "completed":
+    case "success":
+      return "success";
+    case "failed":
+    case "error":
+      return "danger";
+    case "running":
+    case "pending":
+      return "running";
+    case "partial":
+    case "paused":
+    case "stalled":
+      return "info";
+    default:
+      return "neutral";
+  }
+}
+
+function SurfaceRunStatusDot({
+  status,
+  className,
+}: {
+  status: string;
+  className?: string;
+}) {
+  const tone = getRunStatusTone(status);
+
+  return (
+    <span className={cn("inline-flex items-center", className)} title={status}>
+      <span
+        aria-hidden="true"
+        className={cn(
+          "size-2 rounded-full",
+          tone === "success" && "bg-success-9",
+          tone === "danger" && "bg-danger-9",
+          tone === "neutral" && "bg-neutral-7",
+          tone === "info" && "bg-info-9",
+          tone === "running" && "bg-warning-9 animate-[pulse_2s_ease-in-out_infinite]",
+        )}
+      />
+      <span className="sr-only">{status}</span>
+    </span>
+  );
+}
+
+function getSourceVariant(source: RunResponse["source"]) {
+  switch (source) {
+    case "manual":
+      return "neutral";
+    case "webhook":
+      return "info";
+    case "schedule":
+      return "accent";
+    case "simulation":
+      return "warning";
+    default:
+      return "neutral";
+  }
+}
+
+function SurfaceRegressionCell({
+  runId,
+  regressionCount,
+}: {
+  runId: string;
+  regressionCount: number | null | undefined;
+}) {
+  const { data: regressionData } = useRunRegressions(regressionCount ? runId : "");
+
+  if (!regressionCount) {
+    return <span className="text-muted">—</span>;
+  }
+
+  const tooltip = regressionData?.issues?.length
+    ? formatRegressionTooltip(regressionData.issues)
+    : {
+        header: `${regressionCount} ${regressionCount === 1 ? "regression" : "regressions"}`,
+        lines: ["Regression detected"],
+      };
+
+  return (
+    <TooltipProvider>
+      <Tooltip>
+        <TooltipTrigger
+          render={
+            <span className="inline-flex items-center gap-1" style={{ color: "var(--warning-11)" }}>
+              <AlertTriangle className="h-3.5 w-3.5" />
+              {regressionCount}
+            </span>
+          }
+        />
+        <TooltipContent className="max-w-[320px] whitespace-normal px-3 py-3">
+          <RegressionTooltipBody header={tooltip.header} lines={tooltip.lines} />
+        </TooltipContent>
+      </Tooltip>
+    </TooltipProvider>
+  );
+}
+
+function SurfaceRunsTable({
+  runs,
+  currentRunId,
+  onRowClick,
+}: {
+  runs: RunResponse[];
+  currentRunId?: string;
+  onRowClick: (runId: string) => void;
+}) {
+  if (runs.length === 0) {
+    return (
+      <div className="flex h-full items-center justify-center text-sm text-muted">
+        No runs found for this workflow.
+      </div>
+    );
+  }
+
+  return (
+    <div className={SURFACE_RUNS_TABLE_CONTAINER_CLASS}>
+      <Table className={SURFACE_RUNS_TABLE_CLASS}>
+        <TableHeader>
+          <TableRow className={SURFACE_RUNS_TABLE_HEADER_ROW_CLASS}>
+            <TableHead className={cn(SURFACE_RUNS_TABLE_HEAD_CLASS, SURFACE_RUNS_TABLE_STATUS_HEAD_CLASS)}>
+              <span className="sr-only">Status</span>
+            </TableHead>
+            <TableHead className={SURFACE_RUNS_TABLE_HEAD_CLASS}>Run</TableHead>
+            <TableHead className={SURFACE_RUNS_TABLE_HEAD_CLASS}>Source</TableHead>
+            <TableHead className={SURFACE_RUNS_TABLE_HEAD_CLASS}>Started</TableHead>
+            <TableHead className={SURFACE_RUNS_TABLE_HEAD_CLASS}>Duration</TableHead>
+            <TableHead className={SURFACE_RUNS_TABLE_HEAD_CLASS}>Cost</TableHead>
+            <TableHead className={SURFACE_RUNS_TABLE_HEAD_CLASS}>Eval</TableHead>
+            <TableHead className={SURFACE_RUNS_TABLE_HEAD_CLASS}>Regr</TableHead>
+          </TableRow>
+        </TableHeader>
+        <TableBody>
+          {runs.map((run) => {
+            const rowHasRegression = (run.regression_count ?? 0) > 0;
+
+            return (
+              <TableRow
+                key={run.id}
+                className={cn(
+                  SURFACE_RUNS_TABLE_ROW_CLASS,
+                  currentRunId === run.id && "bg-surface-selected",
+                )}
+                onClick={() => onRowClick(run.id)}
+              >
+                <TableCell
+                  className={cn(
+                    SURFACE_RUNS_TABLE_STATUS_CELL_CLASS,
+                    rowHasRegression && SURFACE_RUNS_TABLE_REGRESSION_STRIPE_CLASS,
+                  )}
+                >
+                  <SurfaceRunStatusDot status={run.status} className="w-full justify-center" />
+                </TableCell>
+                <TableCell data-type="data" className={cn(SURFACE_RUNS_TABLE_CELL_CLASS, "text-muted")}>
+                  {run.run_number != null ? `#${run.run_number}` : "—"}
+                </TableCell>
+                <TableCell data-type="data" className={SURFACE_RUNS_TABLE_CELL_CLASS}>
+                  <Badge variant={getSourceVariant(run.source)}>{run.source}</Badge>
+                </TableCell>
+                <TableCell data-type="timestamp" className={cn(SURFACE_RUNS_TABLE_CELL_CLASS, "text-muted")}>
+                  {run.started_at ? getTimeAgo(new Date(run.started_at * 1000).toISOString()) : "—"}
+                </TableCell>
+                <TableCell data-type="metric" className={cn(SURFACE_RUNS_TABLE_CELL_CLASS, "text-secondary")}>
+                  {formatDuration(run.duration_seconds)}
+                </TableCell>
+                <TableCell data-type="metric" className={cn(SURFACE_RUNS_TABLE_CELL_CLASS, "text-success-11")}>
+                  {formatCost(run.total_cost_usd)}
+                </TableCell>
+                <TableCell data-type="metric" className={SURFACE_RUNS_TABLE_CELL_CLASS}>
+                  {typeof run.eval_pass_pct === "number"
+                    ? `${Math.round(run.eval_pass_pct)}%`
+                    : typeof run.eval_score_avg === "number"
+                      ? run.eval_score_avg.toFixed(2)
+                      : <span className="text-muted">—</span>}
+                </TableCell>
+                <TableCell data-type="metric" className={SURFACE_RUNS_TABLE_CELL_CLASS}>
+                  <SurfaceRegressionCell
+                    runId={run.id}
+                    regressionCount={run.regression_count}
+                  />
+                </TableCell>
+              </TableRow>
+            );
+          })}
+        </TableBody>
+      </Table>
+    </div>
+  );
+}
 
 function sseEventToLogEntry(
   eventType: string,
@@ -74,13 +300,13 @@ function sseEventToLogEntry(
   }
 }
 
-function CanvasBottomPanelContent({
+function SurfaceBottomPanelContent({
   runId: initialRunId,
   workflowId,
   defaultState = "collapsed",
   executionSummary,
   regressionsData,
-}: CanvasBottomPanelContentProps) {
+}: SurfaceBottomPanelContentProps) {
   const [isExpanded, setIsExpanded] = useState(defaultState === "expanded");
   const [activeTab, setActiveTab] = useState<"logs" | "runs" | "regressions">("logs");
   const [selectedRunId, setSelectedRunId] = useState<string | undefined>(initialRunId);
@@ -295,7 +521,7 @@ function CanvasBottomPanelContent({
       )}
       {isExpanded && activeTab === "runs" && (
         <div data-testid="workflow-runs-panel" className="overflow-auto flex-1">
-          <RunsTable
+          <SurfaceRunsTable
             runs={sortedRuns}
             currentRunId={currentRunId}
             onRowClick={onRunSelect}
@@ -326,22 +552,22 @@ function CanvasBottomPanelContent({
   );
 }
 
-function RunScopedCanvasBottomPanel(props: CanvasBottomPanelProps & { runId: string }) {
+function RunScopedSurfaceBottomPanel(props: SurfaceBottomPanelProps & { runId: string }) {
   const { data: regressionsData } = useRunRegressions(props.runId);
 
-  return <CanvasBottomPanelContent {...props} regressionsData={regressionsData} />;
+  return <SurfaceBottomPanelContent {...props} regressionsData={regressionsData} />;
 }
 
-function WorkflowScopedCanvasBottomPanel(props: CanvasBottomPanelProps) {
+function WorkflowScopedSurfaceBottomPanel(props: SurfaceBottomPanelProps) {
   const { data: regressionsData } = useWorkflowRegressions(props.workflowId ?? "");
 
-  return <CanvasBottomPanelContent {...props} regressionsData={regressionsData} />;
+  return <SurfaceBottomPanelContent {...props} regressionsData={regressionsData} />;
 }
 
-export function CanvasBottomPanel(props: CanvasBottomPanelProps) {
+export function SurfaceBottomPanel(props: SurfaceBottomPanelProps) {
   if (props.runId) {
-    return <RunScopedCanvasBottomPanel {...props} runId={props.runId} />;
+    return <RunScopedSurfaceBottomPanel {...props} runId={props.runId} />;
   }
 
-  return <WorkflowScopedCanvasBottomPanel {...props} />;
+  return <WorkflowScopedSurfaceBottomPanel {...props} />;
 }
