+++ apps/gui/src/features/surface/SurfaceInspectorPanel.tsx
