+++ apps/gui/src/features/surface/nodes/SoulNode.tsx
