+++ apps/gui/src/features/surface/surfaceUtils.ts
