+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "runsight-workspace"
-version = "0.1.15"
+version = "0.1.16"
 requires-python = ">=3.11"
 dependencies = []
